# app/definition/swagger.yaml
swagger: "2.0"
info:
  version: desp - v1
  title: Despachalo.com
  description: API de servicios de Despachalo.com
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /desp/v1
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /bags:
    get:
      tags:
        - Bolsa
      description: Lista todas las bolsas compradas por el usuario
      # used as the method name of the controller
      operationId: findBags
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: email
          in: query
          description: Correo del usuario
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /bag:
    get:
      tags:
        - Bolsa
      description: Lista la bolsa activa para el usuario
      # used as the method name of the controller
      operationId: findBag
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: email
          in: query
          description: Correo del usuario
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      tags:
        - Bolsa
      description: Registrar bolsa
      # used as the method name of the controller
      operationId: addBag
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/BagAdd"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /addresses:
    get:
      tags:
        - Direccion
      description: Lista las direcciones de usuario
      # used as the method name of the controller
      operationId: findAllAddresses
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: email
          in: query
          description: Correo del usuario
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /address:
    get:
      tags:
        - Direccion
      description: Lista las direcciones de usuario
      # used as the method name of the controller
      operationId: findAddress
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: email
          in: query
          description: Correo del usuario
          required: true
          type: string
        - name: alias
          in: query
          description: Alias de la direccion
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      tags:
        - Direccion
      description: Registrar Direcciones
      # used as the method name of the controller
      operationId: addAddress
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/Address"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      tags:
        - Direccion
      description: Editar Direccion
      # used as the method name of the controller
      operationId: editAddress
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/Address"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags:
        - Direccion
      description: Elimina una direccion
      # used as the method name of the controller
      operationId: deleteAddress
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/AddressDelete"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DespachaloResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  Address:
    type: object
    required:
      - email
      - alias
      - nameCalle
      - number
      - department
      - city
      - commune
      - country
      - nameContact
      - emailContact
      - phoneContact
    properties:
      email:
        description: Correo del usuario
        type: string
      alias:
        description: Nombre de la dirección
        type: string
      nameCalle:
        description: Correo del usuario
        type: string
      number:
        description: Correo del usuario
        type: string
      department:
        description: Correo del usuario
        type: string
      city:
        description: Correo del usuario
        type: string
      commune:
        description: Correo del usuario
        type: string
      country:
        description: Correo del usuario
        type: string
      postalCode:
        description: Correo del usuario
        type: string
      nameContact:
        description: Correo del usuario
        type: string
      emailContact:
        description: Correo del usuario
        type: string
      phoneContact:
        description: Correo del usuario
        type: string
      observations:
        description: Correo del usuario
        type: string
      checkDelivery:
        description: Correo del usuario
        type: string
      checkRetirement:
        description: Correo del usuario
        type: string
  AddressDelete:
    type: object
    required:
      - email
      - alias
    properties:
      email:
        description: Correo del usuario
        type: string
      alias:
        description: Nombre de la dirección
        type: string
...
  BagAdd:
    type: object
    required:
      - email
      - type
    properties:
      email:
        description: Correo del usuario
        type: string
      type:
        description: Tipo de bolsa prepago
        type: string
...
  DespachaloResponse:
    type: string
  ErrorResponse:
    type: string